Refaktorer saksgangvelger (first cut)
 - Fjern ubrukt kode
 - Bytt navn for å tydeliggjøre funksjon til VurderVirksomhetKnapp
 - dra ut hjelpefunksjoner i egne filer
 - Bytt Chevron retning på tilbakestill
 - Bytt navn på Tilbake knapp
